chore(composed-modal): extrapolate out modal header component
